backend/pyproject.toml
[tool.poetry]
name = "msp_server"
version = "0.1.0"
description = "FastAPI와 MySQL을 사용하는 마이크로서비스 프로젝트"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.104.0"
uvicorn = "^0.23.2"
alembic = "^1.12.0"
sqlalchemy = "^2.0.22"
pydantic = "^2.4.2"
pydantic-settings = "^2.0.3"
pymysql = "^1.1.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^23.10.1"
isort = "^5.12.0"
mypy = "^1.6.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88

[tool.isort]
profile = "black"
line_length = 88 